+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # PromptLab Changelog
 
+## [Bug Fixes] - 2023-06-22
+
+- Fixed bug where various commands incorrectly accessed the list of installed commands, causing an error
+- Fixed placeholders for events and reminders not working on macOS Sonoma
+
 ## [Initial Version] - 2023-04-18